filterForm component
