+++ src/components/Slider/Slider.module.css
@@ -1,5 +1,5 @@
 .sliderElement {
-  width: 100%;
+  width: 90%;
   height: auto;
   margin-top: 20px;
   margin-left: 20px;
@@ -56,11 +56,11 @@
 }
 
 .sliderValue {
-  margin: auto;
+  margin: auto 10px;
   padding: 0;
-  width: 40px;
-  height: 40px;
+  width: 30px;
+  height: 30px;
   font-weight: 700;
-  font-size: 40px;
-  line-height: 40px;
+  font-size: 30px;
+  line-height: 30px;
 }
