+++ src/components/TableMain/TableMain.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styles from './TableMain.module.css';
 import TableHeader from '../TableHeader/TableHeader';
-import TableItems from '../TableItem/TableItems';
+import TableItems from '../TableItems/TableItems';
 import { Pagination } from '../Pagination/Pagination';
 import Slider from '../Slider/Slider';
 
@@ -10,33 +10,12 @@ const TableMain = (props) => {
   const [currentPage, setCurrentPage] = useState(1);
   // количество элементов на странице
   const [itemsPerPage, setItemsPerPage] = useState(10);
-  const [maxPriceSliderVal, setMaxPriceSliderVal] = useState(0);
-  const [minPriceSliderVal, setMinPriceSliderVal] = useState(0);
-  const [pricesRange, setPricesRange] = useState({ max: 0, min: 0 });
+
 
   const lastItemIndex = currentPage * itemsPerPage;
   const firsItemIndex = lastItemIndex - itemsPerPage;
   // const sortedItems =  props.data && handleSort(props.data, minPriceSliderVal, maxPriceSliderVal)
-  let currentItemsList =props.data && props.data?.slice(firsItemIndex, lastItemIndex)
-
-
-  function handleSort(array, min, max) {
-    let modArray =[]
-    array.forEach((item) => {
-      if (item.price > min && item.price < max) {
-        modArray.push(item)
-      }
-    })
-    return modArray;
-  }
-
-  // Минимальная и максимальная цена
-  const handleSetPrice = (array) => {
-    const pricesArray = props.data && array.map((item) => item.price);
-    const maxPrice = pricesArray && Math.max(...pricesArray);
-    const minPrice = pricesArray && Math.min(...pricesArray);
-    return { max: maxPrice, min: minPrice }
-  }
+  let currentItemsList = props.data && props.data?.slice(firsItemIndex, lastItemIndex);
 
   useEffect(() => {
     const max = handleSetPrice(props.data).max;
@@ -45,6 +24,16 @@ const TableMain = (props) => {
     setMinPriceSliderVal(min);
     setPricesRange({ max: max, min: min })
   }, [props.data]);
+
+
+
+  // Минимальная и максимальная цена
+  const handleSetPrice = (array) => {
+    const pricesArray = props.data && array.map((item) => item.price);
+    const maxPrice = pricesArray && Math.max(...pricesArray);
+    const minPrice = pricesArray && Math.min(...pricesArray);
+    return { max: maxPrice, min: minPrice }
+  }
 
   return (
     <div className={styles.section}>
@@ -77,29 +66,6 @@ const TableMain = (props) => {
           setValue={setItemsPerPage}
           sliderValue={itemsPerPage}
         />
-        {
-          pricesRange.max !== 0 && pricesRange.min !== 0 &&
-          <>
-            <Slider
-              min={pricesRange.min}
-              max={pricesRange.max}
-              defaultValue={maxPriceSliderVal}
-              title={'Max. price'}
-              setValue={setMaxPriceSliderVal}
-              sliderValue={maxPriceSliderVal}
-            />
-            <Slider
-              min={pricesRange.min}
-              max={pricesRange.max}
-              defaultValue={minPriceSliderVal}
-              title={'Min. price'}
-              setValue={setMinPriceSliderVal}
-              sliderValue={minPriceSliderVal}
-            />
-          </>
-
-
-        }
 
       </div>
